+++ src/Sentry/Laravel/EventHandler.php
@@ -347,7 +347,7 @@ class EventHandler
             Breadcrumb::TYPE_DEFAULT,
             'log.' . $level,
             $message,
-            empty($context) ? [] : ['params' => $context]
+            $context
         ));
     }