+++ src/Sentry/Laravel/SentryHandler.php
@@ -205,7 +205,7 @@ class SentryHandler extends AbstractProcessingHandler
                 if (isset($record['context']['exception']) && $record['context']['exception'] instanceof \Throwable) {
                     $this->hub->captureException($record['context']['exception']);
                 } else {
-                    $this->hub->captureMessage($record['formatted'], $this->getLogLevel($record['level']));
+                    $this->hub->captureMessage($record['message'] ?? $record['formatted'], $this->getLogLevel($record['level']));
                 }
             }
         );