+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## Unreleased
 
+- Remove wrapper around the context of log entry breadcrumbs (#405)
+- Ensure user integrations are always executed after SDK integrations (#474)
+
 ## 2.4.2
 
 - Avoid collision if another package has bound `sentry` in the Laravel container (#467)