+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Remove wrapper around the context of log entry breadcrumbs (#405)
 - Ensure user integrations are always executed after SDK integrations (#474)
+- Fix repeated booted callback registration from performance tracing middleware (#475)
 
 ## 2.4.2