+++ CHANGELOG.md
@@ -2,7 +2,11 @@
 
 ## Unreleased
 
+- Remove wrapper around the context of log entry breadcrumbs (#405)
 - Ensure user integrations are always executed after SDK integrations (#474)
+- Fix repeated booted callback registration from performance tracing middleware (#475)
+- Add tracing support for queue jobs (#478)
+- Add options to disable query and view tracing (#478)
 
 ## 2.4.2