+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 - Add tracing support for queue jobs (#478)
 - Add options to disable query and view tracing (#478)
 - Remove string representation of exception from exceptions logged through log channels (#482)
+- Use message from Monolog record to prevent bloating the log message being recorded with timestamps and log log level (#482)
 
 ## 2.4.2